Rand a code cleanup on the project, removing unnecessary using, adding in readonly modifiers where appropriate, removing unnecessary casts etc.

+++ Gwen.Net.OpenTk/Renderers/OpenTKGL40Renderer.cs
@@ -166,7 +166,7 @@ namespace Gwen.Net.OpenTk.Renderers
                         return;
                     }
 
-                    float dv = (float)delta / (float)oldHeight;
+                    float dv = delta / (float)oldHeight;
 
                     v1 += dv * (v2 - v1);
                 }
@@ -183,7 +183,7 @@ namespace Gwen.Net.OpenTk.Renderers
                         return;
                     }
 
-                    float dv = (float)delta / (float)oldHeight;
+                    float dv = delta / (float)oldHeight;
 
                     v2 -= dv * (v2 - v1);
                 }
@@ -200,7 +200,7 @@ namespace Gwen.Net.OpenTk.Renderers
                         return;
                     }
 
-                    float du = (float)delta / (float)oldWidth;
+                    float du = delta / (float)oldWidth;
 
                     u1 += du * (u2 - u1);
                 }
@@ -217,7 +217,7 @@ namespace Gwen.Net.OpenTk.Renderers
                         return;
                     }
 
-                    float du = (float)delta / (float)oldWidth;
+                    float du = delta / (float)oldWidth;
 
                     u2 -= du * (u2 - u1);
                 }
@@ -295,7 +295,7 @@ namespace Gwen.Net.OpenTk.Renderers
         {
             GL.Viewport(0, 0, width, height);
             GL.UseProgram(shader.Program);
-            GL.Uniform2(shader.Uniforms["uScreenSize"], (float)width, (float)height);
+            GL.Uniform2(shader.Uniforms["uScreenSize"], width, (float)height);
         }
 
         [StructLayout(LayoutKind.Sequential, Pack = 1)]
